chore: move executor shared methods
