+++ packages/nx-go/src/common/index.ts
@@ -1,2 +1,1 @@
-export * from './execution';
 export * from './run';
