+++ packages/nx-go/src/executors/lint/executor.ts
@@ -1,5 +1,5 @@
 import { ExecutorContext } from '@nx/devkit';
-import { executeCommand, extractProjectRoot } from '../../common';
+import { executeCommand, extractProjectRoot } from '../shared';
 import { LintExecutorSchema } from './schema';
 
 /**
