+++ packages/nx-go/src/executors/shared/index.ts
@@ -1,0 +1,1 @@
+export * from './execution';
